requirements.txt
PyAudio
google-cloud-speech
six==1.10.0
google-api-python-client
dialogflow
pyyaml
protobuf-to-dict
